refactor: move stories to dev to fix build
