+++ packages/react/src/Button/Button.features.stories.tsx
@@ -159,41 +159,3 @@ export const LabelWrap = () => {
     </Stack>
   )
 }
-
-export const TrailingCounterAllVariantsVRT = () => {
-  const [count, setCount] = useState(0)
-  const onClick = () => {
-    setCount(count + 1)
-    announce(`Watch ${count + 1}`)
-  }
-  return (
-    <>
-      <Stack direction="vertical" align="start">
-        <Button onClick={onClick} count={count} size="small">
-          Watch
-        </Button>
-        <Button onClick={onClick} disabled count={count}>
-          Watch
-        </Button>
-        <Button onClick={onClick} variant="primary" count={count}>
-          Watch
-        </Button>
-        <Button onClick={onClick} variant="primary" disabled count={count}>
-          Watch
-        </Button>
-        <Button onClick={onClick} variant="danger" count={count}>
-          Watch
-        </Button>
-        <Button onClick={onClick} variant="danger" disabled count={count}>
-          Watch
-        </Button>
-        <Button onClick={onClick} variant="invisible" count={count}>
-          Watch
-        </Button>
-        <Button onClick={onClick} variant="invisible" disabled count={count}>
-          Watch
-        </Button>
-      </Stack>
-    </>
-  )
-}
